Changed the config list type from std::list to std::vector
It makes more sense if the config lists are represented by vectors, because the user will most likely store the possibilities in a vector as he will need random access. Random access is mandatory as he will only get back an index, not the actual value of the list item.